+++ CHANGELOG.md
@@ -11,6 +11,8 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Upgraded stable Flutter analysis SDK to `2.0.4`.
  * Upgraded dartdoc to `0.42.0`.
  * NOTE: Expected reduction in Job-related API calls.
+ * Periodic tasks are now scoped to `global` or `runtimeVersion`.
+   NOTE: added weekly periodic task: `delete-old-neat-task-statuses`.
 
 ## `20210325t074600-all`
  * Tempoarily disabled youtube integration.